==> notebook/2. TMDB.ipynb <==
import pandas as pd
import numpy as np
import requests
from datetime import datetime
import time

import os
from dotenv import load_dotenv 
load_dotenv()

load_dotenv()
tmdb_key = os.getenv("tmdb_key")

load_dotenv()
tmdb_access_token = os.getenv("tmdb_access_token")

bechdel_new_df = pd.read_csv('../data/Bechdel_updated.csv')
bechdel_new_df

# let's check with values i should get from TMDB
movies = pd.read_csv("../data/movies.csv",index_col=0)
movies.info()

# tmbdid, genres, genres, popularity, production_companies, production_countries, release_date, revenue, vote_average  
# vote_count, crew, budget, cast_gender, crew_gender, cast_female_representation, crew_female_representation 

bechdel_id = list(bechdel_new_df["imdbid"])
movies_id = list(movies["imdbid"])
movies_id = [el for el in bechdel_id if el not in movies_id]
len(movies_id)

#movies_id

# change imdbit in the format ddddddd
def get_id(dep_id):
    try:
        dep_id = str(int(dep_id))
        l = len(dep_id)
        new_id = '0'*(7-l)+dep_id
        return new_id
    except:
        pass
    
def change_all_id(lista):
    for i in range(len(lista)):
        lista[i] = get_id(lista[i])
    return lista

movies_id = change_all_id(movies_id)

#movies_id

headers = {
    "accept": "application/json",
    "Authorization": f"Bearer {tmdb_access_token}"
}

credit_url = f"https://api.themoviedb.org/3/movie/tt9059704/credits?language=en-US"
credit_response = requests.get(credit_url, headers=headers).json()
credit_response

cast = credit_response['cast']
crew = credit_response['crew']

details_url = f"https://api.themoviedb.org/3/movie/tt12198524?language=en-US"
response = requests.get(details_url, headers=headers).json()
response

genres = [el["name"] for el in response["genres"]]
imdb_id = response["imdb_id"].replace('tt','')
budget = response["budget"]
popularity = response["popularity"]
production_companies = response['production_companies']
release_date = response['release_date']
revenue = response['revenue']
vote_average = response['vote_average']
vote_count = response['vote_count']

def get_movie_info(imdbid_list):
    headers = {
        "accept": "application/json",
        "Authorization": f"Bearer {tmdb_access_token}"
    }
    
    new_movie_list = []
    cont=0
    
    for imbdid in imdbid_list:
        movie_dict={}
        time.sleep(1)
        details_url = f"https://api.themoviedb.org/3/movie/tt{imbdid}?language=en-US"        
        credits_url = f"https://api.themoviedb.org/3/movie/tt{imbdid}/credits?language=en-US"
        
        
        details_response = requests.get(details_url, headers=headers).json()
        try: 
            movie_dict["imdbid"] = imbdid
            movie_dict["genres"] = [el["name"] for el in details_response["genres"]]           
            movie_dict["budget"] = details_response["budget"]
            movie_dict["popularity"] = details_response["popularity"]
            movie_dict["production_companies"] = details_response['production_companies']
            movie_dict["release_date"] = details_response['release_date']
            movie_dict["revenue"] = details_response['revenue']
            movie_dict["vote_average"] = details_response['vote_average']
            movie_dict["vote_count"] = details_response['vote_count']
        except:
            pass
    
        credit_response = requests.get(credits_url, headers=headers).json()
        
        try:
            movie_dict['cast'] = credit_response['cast']
            movie_dict['crew'] = credit_response['crew']
        except:
            pass
        
        new_movie_list.append(movie_dict)
        
        if cont%20==0:
            pd.DataFrame(new_movie_list).to_csv("../data/checkpoint_df.csv",index=False)
        cont+=1
        
    return new_movie_list

movie_list = get_movie_info(movies_id)

df = pd.read_csv("../data/checkpoint_df.csv")
df

#df = pd.DataFrame(movie_list)
df.dropna(inplace=True)
df.drop_duplicates(inplace=True)
df.reset_index(drop=True,inplace=True)

df

# junta los 3 dataframes y elimina columnas duplicadas
# bechdel_new_df = pd.read_csv("../data/Bechdel_updated.csv")

new_movies_df = pd.merge(bechdel_new_df, df, left_on='imdbid', right_on='imdbid', how='inner')

new_movies_df.info()

new_movies_df.drop(columns=["id","submitterid","date","visible"], inplace=True)

new_movies_df.info()

# function to get cast/crew gender

for index, movie in new_movies_df.iterrows():
    companies = eval(movie['production_companies'])
    companies_list = []

    for company in companies:
        company_name = company['name']
        companies_list.append(company_name)

    new_movies_df.at[index, 'production_companies'] = companies_list

for index, movie in new_movies_df.iterrows():
    date = datetime.strptime(movie['release_date'], '%Y-%m-%d')
    reformatted_date = date.strftime('%d/%m/%Y')

    new_movies_df.at[index, 'release_date'] = reformatted_date

new_movies_df['cast_gender'] = pd.NA


for index, movie in new_movies_df.iterrows():
    cast = eval(movie['cast'])
    gender_list = []

    for cast_member in cast:
        gender = cast_member['gender']
        gender_list.append(gender)

    new_movies_df.at[index, 'cast_gender'] = gender_list

new_movies_df['crew_gender'] = pd.NA


for index, movie in new_movies_df.iterrows():
    crew = eval(movie['crew'])
    gender_list = []

    for crew_member in crew:
        gender = crew_member['gender']
        gender_list.append(gender)

    new_movies_df.at[index, 'crew_gender'] = gender_list

new_movies_df['cast_female_representation'] = pd.NA
new_movies_df['crew_female_representation'] = pd.NA


for index, movie in new_movies_df.iterrows():
    if len(movie['cast_gender']) != 0:
        new_movies_df.at[index, 'cast_female_representation'] = 100*(movie['cast_gender'].count(1)/len(movie['cast_gender']))

    if len(movie['crew_gender']) != 0:
        new_movies_df.at[index, 'crew_female_representation'] = 100*(movie['crew_gender'].count(1)/len(movie['crew_gender']))


new_movies_df = new_movies_df.dropna().copy()
new_movies_df['cast_female_representation'] = new_movies_df['cast_female_representation'].astype('float64')
new_movies_df['crew_female_representation'] = new_movies_df['crew_female_representation'].astype('float64')

movies = pd.read_csv("../data/movies.csv", index_col=0)
movies.info()

new_movies_df.info()

all_movies = pd.concat([movies,new_movies_df], axis=0)
all_movies.info()

all_movies.dropna(how='all',inplace=True)

all_movies.drop(columns=["cast","crew"],inplace=True)

all_movies.drop_duplicates(subset="imdbid", inplace=True)

all_movies.info()

all_movies.reset_index(drop=True, inplace=True)

all_movies

all_movies.to_csv("../data/all_movies.csv")
