==> notebook/3. exploring.ipynb <==
import pandas as pd
import numpy as np
import ast

import pymysql
import sqlalchemy as alch

import os
from dotenv import load_dotenv 
load_dotenv()

sql_password = os.getenv("SQL_password")

movies = pd.read_csv("../data/movies.csv", index_col=0)

movies.info()

movies.sample(4)

- **bt_score** : The bechdel test score. <br>
    0: Failed the test :( <br>
    1: It has to have at least two named women in it.<br>
    2: Who talk to each other.<br>
    3: About something besides a man.
    
- **dubious**: Whether the submitter considered the rating dubious.

- **imdbid**: id on IMDb 

- **tmdbId**: id on TMDB

- **popularity**: TMDB popularity.

- **revenue**: revenue in American Dollar (USD). Not adjusted for inflation.

- **vote_average**: average rating

- **vote_count**: number of ratings
    
- **cast and crew**: a json with cast and crew information. Check the response body in the TMDB api docs.

- **cast_gender and crew_gender**: gender of cast and crew.<br>
    0: Not specified;<br>
    1: Female;<br>
    2: Male;<br>
    3: Non-binary.
    
- **cast_female_representation** and **crew_female_representation**: percentage of women (1) in the cast or crew.

# nuove colonne: regista, produttore

#set(movies["release_date"])

movies[~movies["release_date"].str.contains(r'\d{2}/\d{2}/\d{4}')]["release_date"]

movies.drop(columns=['crew','cast'], inplace=True)

all_movies=pd.read_csv("../data/all_movies.csv", index_col=0)
all_movies

all_movies.info()

#dataframe per producing companies
'''
imbdid (pk)
title
release_year
producing_company
budget
bt_score       
dubious 
'''

type(all_movies["production_companies"][0])

#creating a dataframe for production companies
production_companies_lst=[]
for index, row in all_movies.iterrows():
    # "production_companies" is stored as a string representation of a list, i had to convert it into a list first
    production_companies_list = ast.literal_eval(row["production_companies"])
    for el in production_companies_list:
        movie_dict = {}
        movie_dict["imdbid"] = row["imdbid"]
        movie_dict["title"] = row["title"]
        #movie_dict["year"] = row["year"]
        movie_dict["production_company"] = el
        #movie_dict["bt_score"] = row["bt_score"]
        #movie_dict["dubious"] = row["dubious"]
        #movie_dict["budget"] = row["budget"]
        production_companies_lst.append(movie_dict)

production_companies = pd.DataFrame(production_companies_lst)

production_companies.sample(10)

production_companies.to_csv("../data/production_companies.csv", index=False)

# dataframe per cast gender

cast_lst=[]
for index, row in all_movies.iterrows():
    movie_dict = {}
    movie_dict["imdbid"] = row["imdbid"]
    movie_dict["title"] = row["title"]
    #movie_dict["bt_score"] = row["bt_score"]
    movie_dict["female_cast_count"]=0
    movie_dict["male_cast_count"]=0
    movie_dict["nonbinary_cast_count"]=0
    cast_gender_lst = ast.literal_eval(row["cast_gender"])
    for el in cast_gender_lst:
        if el==1:
            movie_dict["female_cast_count"]+=1
        elif el==2:
            movie_dict["male_cast_count"]+=1
        elif el==3:
            movie_dict["nonbinary_cast_count"]+=1
    movie_dict["num_actors"]=len(cast_gender_lst)
    cast_lst.append(movie_dict)

cast = pd.DataFrame(cast_lst)

cast.to_csv("../data/cast.csv", index=False)

#dataframe per crew gender
crew_lst=[]
for index, row in all_movies.iterrows():
    movie_dict = {}
    movie_dict["imdbid"] = row["imdbid"]
    movie_dict["title"] = row["title"]
    #movie_dict["bt_score"] = row["bt_score"]
    movie_dict["female_crew"]=0
    movie_dict["male_crew"]=0
    movie_dict["nonbinary_crew"]=0
    crew_gender_lst = ast.literal_eval(row["crew_gender"])
    for el in crew_gender_lst:
        if el==1:
            movie_dict["female_crew"]+=1
        elif el==2:
            movie_dict["male_crew"]+=1
        elif el==3:
            movie_dict["nonbinary_crew"]+=1
    movie_dict["num_crew"]=len(crew_gender_lst)
    crew_lst.append(movie_dict)

crew = pd.DataFrame(crew_lst)

crew

crew.to_csv("../data/crew.csv", index=False)

# SAVE TO SQL

dbName = "women_representation"

connectionData = f"mysql+pymysql://root:{sql_password}@localhost/{dbName}"

engine = alch.create_engine(connectionData)

all_movies.to_sql("movies", if_exists="append", con=engine, index=False)

production_companies.to_sql("production_companies", if_exists="append", con=engine, index=False)

cast.to_sql("cast", if_exists="append", con=engine, index=False)

crew.to_sql("crew", if_exists="append", con=engine, index=False)

engine.execute('''
ALTER TABLE movies
ADD PRIMARY KEY (imdbid);
''')

engine.execute('''
ALTER TABLE cast
ADD PRIMARY KEY (imdbid);
''')

engine.execute( '''
ALTER TABLE crew
ADD PRIMARY KEY (imdbid);
''');

engine.execute('''
ALTER TABLE production_companies
ADD FOREIGN KEY (imdbid)
REFERENCES movies(imdbid)
''');

engine.execute('''
ALTER TABLE crew
ADD FOREIGN KEY (imdbid)
REFERENCES movies(imdbid)
''');

engine.execute('''
ALTER TABLE cast
ADD FOREIGN KEY (imdbid)
REFERENCES movies(imdbid)
''');
