==> notebook/4. SQL queries.ipynb <==
import pymysql
import pandas as pd
import sqlalchemy as alch

import os
from dotenv import load_dotenv 
load_dotenv()

sql_password = os.getenv("SQL_password")

dbName = "women_representation"
connectionData=f"mysql+pymysql://root:{sql_password}@localhost/{dbName}"
engine = alch.create_engine(connectionData)

query = '''

'''

pd.read_sql_query(query, engine)

# 1
#percentage of each b_score
#also depending on year

# 2
# best producing companies

# 3
#relation btw b_score and incassi/votazione

# 4
#correlation btw b_score and cast/crew gender

#1

# percentuale di film per ogni bechdel score
query = '''
SELECT bt_score AS Bechdel_score,
       COUNT(*) * 100.0 / (SELECT COUNT(*) FROM movies) AS percentage
FROM movies
GROUP BY bt_score
'''

pd.read_sql_query(query, engine)

# percentuale di film nel corso degli anni per ogni bechdel score
query = '''
SELECT 
    release_year,
    SUM(CASE WHEN bt_score = 0 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt0_percentage,
    SUM(CASE WHEN bt_score = 1 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt1_percentage,
    SUM(CASE WHEN bt_score = 2 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt2_percentage,
    SUM(CASE WHEN bt_score = 3 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt3_percentage,
    COUNT(*) AS num_movies
FROM (SELECT title, bt_score, YEAR(STR_TO_DATE(release_date, '%%d/%%m/%%Y')) AS release_year
FROM movies) AS tab
GROUP BY release_year
ORDER BY release_year DESC;
'''

pd.read_sql_query(query, engine)

# percentuale di film nel corso dei decenni per ogni bechdel score
query = '''
SELECT 
    FLOOR(release_year / 10) * 10 AS decade,
    SUM(CASE WHEN bt_score = 0 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt0_percentage,
    SUM(CASE WHEN bt_score = 1 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt1_percentage,
    SUM(CASE WHEN bt_score = 2 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt2_percentage,
    SUM(CASE WHEN bt_score = 3 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt3_percentage,
    COUNT(*) AS num_movies
FROM (
    SELECT 
        title, 
        bt_score, 
        FLOOR(YEAR(STR_TO_DATE(release_date, '%%d/%%m/%%Y')) / 10) * 10 AS release_year
    FROM movies
) AS tab
GROUP BY decade
HAVING num_movies>100
ORDER BY decade DESC;
'''

pd.read_sql_query(query, engine)

#2

# percentuale di film per producing companies per ogni bechdel score
query = '''
SELECT 
    production_company,
    SUM(CASE WHEN bt_score = 0 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt0_percentage,
    SUM(CASE WHEN bt_score = 1 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt1_percentage,
    SUM(CASE WHEN bt_score = 2 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt2_percentage,
    SUM(CASE WHEN bt_score = 3 THEN 1 ELSE 0 END) / COUNT(*) * 100 AS bt3_percentage,
    COUNT(*) AS num_movies
FROM (SELECT production_companies.title, bt_score, production_company
FROM production_companies
JOIN movies ON production_companies.imdbid = movies.imdbid) AS tab
GROUP BY production_company
HAVING num_movies > 60
ORDER BY num_movies DESC;
'''

pd.read_sql_query(query, engine)

# producing company - average budget - average budget b_score=3
query = '''
SELECT 
    production_company,
    AVG(budget)/1000000 AS avg_Mbudget,
    AVG(CASE WHEN bt_score = 3 THEN budget ELSE NULL END)/1000000 AS avg_Mbudget_bt3,
    COUNT(*) AS num_movies
FROM (SELECT production_companies.title, bt_score, production_company, budget
FROM production_companies
JOIN movies ON production_companies.imdbid = movies.imdbid
WHERE budget>0) AS tab
GROUP BY production_company
HAVING num_movies > 50
ORDER BY avg_Mbudget_bt3-avg_Mbudget DESC; 
'''

# ordino mettendo per prime le producing company con maggiore differenza tra budget per bt_score3 e budget 
pd.read_sql_query(query, engine)

# 3

# Average of film popularity for each bechdel test value
query = '''
SELECT bt_score AS Bechdel_score, AVG(popularity) AS Average_Popularity
FROM movies
GROUP BY bt_score

'''

pd.read_sql_query(query, engine)

# Average of revenues for each bechdel test value
query = '''
SELECT bt_score AS Bechdel_score, AVG(revenue)/1000000 AS Average_M_Revenue
FROM movies
GROUP BY bt_score

'''

pd.read_sql_query(query, engine)

# 4

# correlation btw b_score and cast gender
query = '''
SELECT     (
        COUNT(*) * SUM(female_ratio * bt_score) - SUM(female_ratio) * SUM(bt_score)
    ) / SQRT(
        (COUNT(*) * SUM(female_ratio * bt_score) - SUM(female_ratio) * SUM(female_ratio)) *
        (COUNT(*) * SUM(bt_score * bt_score) - SUM(bt_score) * SUM(bt_score))
    ) AS correlation
FROM 
(SELECT cast.title, female_cast_count/num_actors AS female_ratio, bt_score FROM cast
JOIN movies on cast.imdbid=movies.imdbid) AS tab

'''

pd.read_sql_query(query, engine)

#correlation btw b_score and crew gender
query = '''
SELECT     (
        COUNT(*) * SUM(female_ratio * bt_score) - SUM(female_ratio) * SUM(bt_score)
    ) / SQRT(
        (COUNT(*) * SUM(female_ratio * bt_score) - SUM(female_ratio) * SUM(female_ratio)) *
        (COUNT(*) * SUM(bt_score * bt_score) - SUM(bt_score) * SUM(bt_score))
    ) AS correlation
FROM 
(SELECT crew.title, female_crew/num_crew AS female_ratio, bt_score FROM crew
JOIN movies on crew.imdbid=movies.imdbid) AS tab

'''

pd.read_sql_query(query, engine)

#
query = '''
SELECT bt_score AS Bechdel_score,
       AVG(female_ratio) AS avg_fem_ratio,
       AVG(male_ratio) AS avg_mal_ration,
       AVG(nonbinary_ratio) AS avg_nbin_ratio       
FROM (
SELECT crew.title, bt_score, female_crew/num_crew AS female_ratio, male_crew/num_crew AS male_ratio,nonbinary_crew/num_crew AS nonbinary_ratio
FROM crew
JOIN movies on crew.imdbid=movies.imdbid) AS tab
GROUP BY bt_score
'''

pd.read_sql_query(query, engine)

 # non si arriva a 100 perche c'erano anche valori unknown
